[Core] Improve fixtures, extract example factories for further usage
