+++ src/Sylius/Bundle/CoreBundle/Fixture/Factory/TaxCategoryExampleFactory.php
@@ -1,0 +1,78 @@
+<?php
+
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Paweł Jędrzejewski
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sylius\Bundle\CoreBundle\Fixture\Factory;
+
+use Sylius\Component\Core\Formatter\StringInflector;
+use Sylius\Component\Resource\Factory\FactoryInterface;
+use Sylius\Component\Taxation\Model\TaxCategoryInterface;
+use Symfony\Component\OptionsResolver\Options;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+/**
+ * @author Kamil Kokot <kamil.kokot@lakion.com>
+ */
+final class TaxCategoryExampleFactory implements ExampleFactoryInterface
+{
+    /**
+     * @var FactoryInterface
+     */
+    private $taxCategoryFactory;
+
+    /**
+     * @var \Faker\Generator
+     */
+    private $faker;
+
+    /**
+     * @var OptionsResolver
+     */
+    private $optionsResolver;
+
+    /**
+     * @param FactoryInterface $taxCategoryFactory
+     */
+    public function __construct(FactoryInterface $taxCategoryFactory)
+    {
+        $this->taxCategoryFactory = $taxCategoryFactory;
+
+        $this->faker = \Faker\Factory::create();
+        $this->optionsResolver =
+            (new OptionsResolver())
+                ->setDefault('name', function (Options $options) {
+                    return $this->faker->words(3, true);
+                })
+                ->setDefault('code', function (Options $options) {
+                    return StringInflector::nameToCode($options['name']);
+                })
+                ->setDefault('description', function (Options $options) {
+                    return $this->faker->paragraph;
+                })
+        ;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function create(array $options = [])
+    {
+        $options = $this->optionsResolver->resolve($options);
+        
+        /** @var TaxCategoryInterface $taxCategory */
+        $taxCategory = $this->taxCategoryFactory->createNew();
+
+        $taxCategory->setCode($options['code']);
+        $taxCategory->setName($options['name']);
+        $taxCategory->setDescription($options['description']);
+
+        return $taxCategory;
+    }
+}
