+++ src/Sylius/Bundle/CoreBundle/Fixture/ShippingCategoryFixture.php
@@ -11,45 +11,13 @@
 
 namespace Sylius\Bundle\CoreBundle\Fixture;
 
-use Doctrine\Common\Persistence\ObjectManager;
-use Sylius\Bundle\FixturesBundle\Fixture\AbstractFixture;
-use Sylius\Component\Core\Formatter\StringInflector;
-use Sylius\Component\Resource\Factory\FactoryInterface;
-use Sylius\Component\Shipping\Model\ShippingCategoryInterface;
 use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
-use Symfony\Component\OptionsResolver\Options;
-use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
  * @author Kamil Kokot <kamil.kokot@lakion.com>
  */
 final class ShippingCategoryFixture extends AbstractResourceFixture
 {
-    /**
-     * @var FactoryInterface
-     */
-    private $shippingCategoryFactory;
-
-    /**
-     * @var \Faker\Generator
-     */
-    private $faker;
-
-    /**
-     * @param FactoryInterface $shippingCategoryFactory
-     * @param ObjectManager $shippingCategoryManager
-     */
-    public function __construct(
-        FactoryInterface $shippingCategoryFactory,
-        ObjectManager $shippingCategoryManager
-    ) {
-        parent::__construct($shippingCategoryManager, 'shipping_categories', 'name');
-
-        $this->shippingCategoryFactory = $shippingCategoryFactory;
-
-        $this->faker = \Faker\Factory::create();
-    }
-
     /**
      * {@inheritdoc}
      */
@@ -61,56 +29,13 @@ final class ShippingCategoryFixture extends AbstractResourceFixture
     /**
      * {@inheritdoc}
      */
-    protected function loadResource(array $options)
-    {
-        /** @var ShippingCategoryInterface $shippingCategory */
-        $shippingCategory = $this->shippingCategoryFactory->createNew();
-
-        $shippingCategory->setCode($options['code']);
-        $shippingCategory->setName($options['name']);
-        $shippingCategory->setDescription($options['description']);
-
-        return $shippingCategory;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     protected function configureResourceNode(ArrayNodeDefinition $resourceNode)
     {
         $resourceNode
             ->children()
-            ->scalarNode('code')->cannotBeEmpty()->end()
-            ->scalarNode('description')->cannotBeEmpty()->end()
+                ->scalarNode('name')->cannotBeEmpty()->end()
+                ->scalarNode('code')->cannotBeEmpty()->end()
+                ->scalarNode('description')->cannotBeEmpty()->end()
         ;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function configureResourceOptionsResolver(array $options, OptionsResolver $optionsResolver)
-    {
-        $optionsResolver
-            ->setRequired('name')
-            ->setDefault('code', function (Options $options) {
-                return StringInflector::nameToCode($options['name']);
-            })
-            ->setDefault('description', function (Options $options) {
-                return $this->faker->paragraph;
-            })
-        ;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function generateResourcesOptions($amount)
-    {
-        $resourcesOptions = [];
-        for ($i = 0; $i < $amount; ++$i) {
-            $resourcesOptions[] = ['name' => $this->faker->words(3, true)];
-        }
-
-        return $resourcesOptions;
     }
 }
