+++ src/Sylius/Bundle/CoreBundle/Fixture/TaxCategoryFixture.php
@@ -11,44 +11,13 @@
 
 namespace Sylius\Bundle\CoreBundle\Fixture;
 
-use Doctrine\Common\Persistence\ObjectManager;
-use Sylius\Component\Core\Formatter\StringInflector;
-use Sylius\Component\Resource\Factory\FactoryInterface;
-use Sylius\Component\Taxation\Model\TaxCategoryInterface;
 use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
-use Symfony\Component\OptionsResolver\Options;
-use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
  * @author Kamil Kokot <kamil.kokot@lakion.com>
  */
 final class TaxCategoryFixture extends AbstractResourceFixture
 {
-    /**
-     * @var FactoryInterface
-     */
-    private $taxCategoryFactory;
-
-    /**
-     * @var \Faker\Generator
-     */
-    private $faker;
-
-    /**
-     * @param FactoryInterface $taxCategoryFactory
-     * @param ObjectManager $taxCategoryManager
-     */
-    public function __construct(
-        FactoryInterface $taxCategoryFactory,
-        ObjectManager $taxCategoryManager
-    ) {
-        parent::__construct($taxCategoryManager, 'tax_categories', 'name');
-
-        $this->taxCategoryFactory = $taxCategoryFactory;
-
-        $this->faker = \Faker\Factory::create();
-    }
-
     /**
      * {@inheritdoc}
      */
@@ -60,56 +29,13 @@ final class TaxCategoryFixture extends AbstractResourceFixture
     /**
      * {@inheritdoc}
      */
-    protected function loadResource(array $options)
-    {
-        /** @var TaxCategoryInterface $taxCategory */
-        $taxCategory = $this->taxCategoryFactory->createNew();
-
-        $taxCategory->setCode($options['code']);
-        $taxCategory->setName($options['name']);
-        $taxCategory->setDescription($options['description']);
-
-        return $taxCategory;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     protected function configureResourceNode(ArrayNodeDefinition $resourceNode)
     {
         $resourceNode
             ->children()
-            ->scalarNode('code')->cannotBeEmpty()->end()
-            ->scalarNode('description')->cannotBeEmpty()->end()
+                ->scalarNode('name')->cannotBeEmpty()->end()
+                ->scalarNode('code')->cannotBeEmpty()->end()
+                ->scalarNode('description')->cannotBeEmpty()->end()
         ;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function configureResourceOptionsResolver(array $options, OptionsResolver $optionsResolver)
-    {
-        $optionsResolver
-            ->setRequired('name')
-            ->setDefault('code', function (Options $options) {
-                return StringInflector::nameToCode($options['name']);
-            })
-            ->setDefault('description', function (Options $options) {
-                return $this->faker->paragraph;
-            })
-        ;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function generateResourcesOptions($amount)
-    {
-        $resourcesOptions = [];
-        for ($i = 0; $i < $amount; ++$i) {
-            $resourcesOptions[] = ['name' => $this->faker->words(3, true)];
-        }
-
-        return $resourcesOptions;
     }
 }
