+++ src/Sylius/Bundle/CoreBundle/Resources/config/fixtures_factories.xml
@@ -1,0 +1,94 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!--
+
+ This file is part of the Sylius package.
+
+ (c) Paweł Jędrzejewski
+
+ For the full copyright and license information, please view the LICENSE
+ file that was distributed with this source code.
+
+-->
+
+<container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
+    <services>
+        <service id="sylius.fixture.example_factory.payment_method" class="Sylius\Bundle\CoreBundle\Fixture\Factory\PaymentMethodExampleFactory">
+            <argument type="service" id="sylius.factory.payment_method" />
+            <argument type="service" id="sylius.repository.locale" />
+        </service>
+
+        <service id="sylius.fixture.example_factory.shipping_category" class="Sylius\Bundle\CoreBundle\Fixture\Factory\ShippingCategoryExampleFactory">
+            <argument type="service" id="sylius.factory.shipping_category" />
+        </service>
+
+        <service id="sylius.fixture.example_factory.shipping_method" class="Sylius\Bundle\CoreBundle\Fixture\Factory\ShippingMethodExampleFactory">
+            <argument type="service" id="sylius.factory.shipping_method" />
+            <argument type="service" id="sylius.repository.zone" />
+            <argument type="service" id="sylius.repository.shipping_category" />
+            <argument type="service" id="sylius.repository.locale" />
+        </service>
+
+        <service id="sylius.fixture.example_factory.channel" class="Sylius\Bundle\CoreBundle\Fixture\Factory\ChannelExampleFactory">
+            <argument type="service" id="sylius.factory.channel" />
+            <argument type="service" id="sylius.repository.locale" />
+            <argument type="service" id="sylius.repository.currency" />
+            <argument type="service" id="sylius.repository.payment_method" />
+            <argument type="service" id="sylius.repository.shipping_method" />
+        </service>
+
+        <service id="sylius.fixture.example_factory.user" class="Sylius\Bundle\CoreBundle\Fixture\Factory\UserExampleFactory">
+            <argument type="service" id="sylius.factory.user" />
+            <argument type="service" id="sylius.factory.customer" />
+            <argument type="service" id="sylius.repository.currency" />
+        </service>
+
+        <service id="sylius.fixture.example_factory.product_option" class="Sylius\Bundle\CoreBundle\Fixture\Factory\ProductOptionExampleFactory">
+            <argument type="service" id="sylius.factory.product_option" />
+            <argument type="service" id="sylius.factory.product_option_value" />
+            <argument type="service" id="sylius.repository.locale" />
+        </service>
+
+        <service id="sylius.fixture.example_factory.product_attribute" class="Sylius\Bundle\CoreBundle\Fixture\Factory\ProductAttributeExampleFactory">
+            <argument type="service" id="sylius.factory.product_attribute" />
+            <argument type="service" id="sylius.repository.locale" />
+            <argument>%sylius.attribute.attribute_types%</argument>
+        </service>
+
+        <service id="sylius.fixture.example_factory.product_archetype" class="Sylius\Bundle\CoreBundle\Fixture\Factory\ProductArchetypeExampleFactory">
+            <argument type="service" id="sylius.factory.product_archetype" />
+            <argument type="service" id="sylius.repository.product_option" />
+            <argument type="service" id="sylius.repository.product_attribute" />
+            <argument type="service" id="sylius.repository.locale" />
+        </service>
+
+        <service id="sylius.fixture.example_factory.tax_category" class="Sylius\Bundle\CoreBundle\Fixture\Factory\TaxCategoryExampleFactory">
+            <argument type="service" id="sylius.factory.tax_category" />
+        </service>
+
+        <service id="sylius.fixture.example_factory.tax_rate" class="Sylius\Bundle\CoreBundle\Fixture\Factory\TaxRateExampleFactory">
+            <argument type="service" id="sylius.factory.tax_rate" />
+            <argument type="service" id="sylius.repository.zone" />
+            <argument type="service" id="sylius.repository.tax_category" />
+        </service>
+
+        <service id="sylius.fixture.example_factory.taxon" class="Sylius\Bundle\CoreBundle\Fixture\Factory\TaxonExampleFactory">
+            <argument type="service" id="sylius.factory.taxon" />
+            <argument type="service" id="sylius.manager.taxon" />
+            <argument type="service" id="sylius.repository.taxon" />
+            <argument type="service" id="sylius.repository.locale" />
+        </service>
+
+        <service id="sylius.fixture.example_factory.product" class="Sylius\Bundle\CoreBundle\Fixture\Factory\ProductExampleFactory">
+            <argument type="service" id="sylius.factory.product" />
+            <argument type="service" id="sylius.generator.variant" />
+            <argument type="service" id="sylius.factory.product_attribute_value" />
+            <argument type="service" id="sylius.repository.taxon" />
+            <argument type="service" id="sylius.repository.product_archetype" />
+            <argument type="service" id="sylius.repository.product_option" />
+            <argument type="service" id="sylius.repository.shipping_category" />
+            <argument type="service" id="sylius.repository.channel" />
+            <argument type="service" id="sylius.repository.locale" />
+        </service>
+    </services>
+</container>
