+++ src/Sylius/Bundle/CoreBundle/Tests/Fixture/GeographicalFixtureTest.php
@@ -15,14 +15,13 @@ use Doctrine\Common\Persistence\ObjectManager;
 use Matthias\SymfonyConfigTest\PhpUnit\ConfigurationTestCaseTrait;
 use Sylius\Bundle\AddressingBundle\Factory\ZoneFactoryInterface;
 use Sylius\Bundle\CoreBundle\Fixture\GeographicalFixture;
-use Sylius\Component\Addressing\Model\ZoneInterface;
 use Sylius\Component\Resource\Factory\FactoryInterface;
 use Symfony\Component\Intl\Intl;
 
 /**
  * @author Kamil Kokot <kamil.kokot@lakion.com>
  */
-final class GeographicalFixtureConfigurationTest extends \PHPUnit_Framework_TestCase
+final class GeographicalFixtureTest extends \PHPUnit_Framework_TestCase
 {
     use ConfigurationTestCaseTrait;
